fix: [ANDROAPP-5413] Review all ignored tests in TeiDashboardTest file (#3250)
* fix: [ANDROAPP-5413] shouldSuccessfullyScheduleAnEvent test, ignore remove from other tests

* fix: [ANDROAPP] shouldDeleteTeiSuccessfully and shouldDeleteEnrollmentSuccessfully moved to TeiDashboardTestNoComposable, name and lastName modified, search modified

* fix: [ANDROAPP-5413] shouldMakeAReferral fixed

* [ANDROAPP-5418] Remove crashlyticsController from timber (#3241)

* [ANDROAPP-5403] post closeKeyboard (#3235)

* fix deleting values in stock use case (#3236)

* activate desugaring to move zxing-android-embedded from jcenter(). jCenter() is completely removed from project (#3246)

* [ANDROAPP-5413] comments removed

* [ANDROAPP-5340] parse resource value uid to patch before rendering (#3249)

* [ANDROAPP-5426] DataSet tables not saving text values (#3248)

* fix: [ANDROAPP-5328] reset to device language on logout (#3245)

* removed: [ANDROAPP-5413] comments deleted, unused code deleted

---------
