+++ app/src/androidTest/java/org/dhis2/usescases/teidashboard/TeiDashboardTestNoComposable.kt
@@ -7,6 +7,7 @@ import org.dhis2.usescases.searchTrackEntity.SearchTEActivity
 import org.dhis2.usescases.searchte.robot.searchTeiRobot
 import org.dhis2.usescases.teidashboard.robot.relationshipRobot
 import org.dhis2.usescases.teidashboard.robot.teiDashboardRobot
+import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -17,6 +18,7 @@ class TeiDashboardTestNoComposable : BaseTest() {
     @get:Rule
     val ruleSearch = ActivityTestRule(SearchTEActivity::class.java, false, false)
 
+    @Ignore
     @Test
     fun shouldSuccessfullyCreateRelationshipWhenClickAdd() {
         val teiName = "Tim"
@@ -54,6 +56,64 @@ class TeiDashboardTestNoComposable : BaseTest() {
 
         relationshipRobot {
             checkRelationshipWasCreated(0, completeName)
+        }
+    }
+
+    @Test
+    fun shouldDeleteTeiSuccessfully() {
+        val teiName = "Gertrude"
+        val teiLastName = "Fjordsen"
+        val firstNamePosition = 0
+        val lastNamePosition = 1
+
+        setupCredentials()
+        prepareChildProgrammeIntentAndLaunchActivity(ruleSearch)
+
+        searchTeiRobot {
+            clickOnOpenSearch()
+            typeAttributeAtPosition(teiName, firstNamePosition)
+            typeAttributeAtPosition(teiLastName, lastNamePosition)
+            clickOnSearch()
+            clickOnTEI(teiName, teiLastName)
+            //scrollToTEIandClick()
+        }
+
+        teiDashboardRobot {
+            clickOnMenuMoreOptions()
+            clickOnMenuDeleteTEI()
+        }
+
+        searchTeiRobot {
+            checkTEIsDelete(teiName, teiLastName)
+        }
+    }
+
+    @Test
+    fun shouldDeleteEnrollmentSuccessfully() {
+        val teiName = "Anna"
+        val teiLastName = "Jones"
+        val firstNamePosition = 0
+        val lastNamePosition = 1
+
+        setupCredentials()
+        prepareChildProgrammeIntentAndLaunchActivity(ruleSearch)
+
+        searchTeiRobot {
+            clickOnOpenSearch()
+            typeAttributeAtPosition(teiName, firstNamePosition)
+            typeAttributeAtPosition(teiLastName, lastNamePosition)
+            clickOnSearch()
+            //     waitToDebounce(400)
+            clickOnTEI(teiName, teiLastName)
+        }
+
+        teiDashboardRobot {
+            clickOnMenuMoreOptions()
+            clickOnMenuDeleteEnrollment()
+        }
+
+        searchTeiRobot {
+            checkTEIsDelete(teiName, teiLastName)
         }
     }
 }
